refactor: use fetchSpecFromUrl helper only (wip)
