+++ packages/swagger-editor/src/helpers/fetchSpecFromUrl.ts
@@ -1,11 +1,11 @@
-export const fetchSpecFromUrl = async (
-  url: string,
-  { proxyUrl }: { proxyUrl?: string },
-) => {
+/**
+ * Fetches a spec file from a given URL.
+ */
+export const fetchSpecFromUrl = async (url: string, proxy: string) => {
   // With Proxy
-  if (proxyUrl) {
-    const response = proxyUrl
-      ? await fetch(proxyUrl, {
+  if (proxy) {
+    const response = proxy
+      ? await fetch(proxy, {
           method: 'POST',
           headers: {
             'Content-Type': 'application/json',
@@ -30,8 +30,8 @@ export const fetchSpecFromUrl = async (
       )
     }
 
+    // TODO: Deal with Yaml
     const data = JSON.parse(payload.data)
-
     return JSON.stringify(data, null, 2)
   }
   // Without proxy
@@ -40,6 +40,8 @@ export const fetchSpecFromUrl = async (
       '[fetchSpecFromUrl] Trying to fetch the spec file without a proxy. The CORS headers have to be set properly, otherwise the request will fail.',
     )
     const response = await fetch(url)
+
+    // TODO: Deal with Yaml
     const json = await response.json()
     return JSON.stringify(json, null, 2)
   }
