+++ packages/swagger-editor/src/index.ts
@@ -2,6 +2,7 @@ export {
   SwaggerEditor,
   SwaggerEditorGettingStarted,
 } from './components/SwaggerEditor'
+export * from './helpers'
 export * from './hooks'
 export * from './types'
